backend/working-app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
pydantic>=2.0.0
gitpython==3.1.43
python-dotenv==1.0.1
httpx>=0.27.0
aiofiles>=24.1.0
python-multipart>=0.0.9
requests>=2.32.0
numpy>=1.26.0
reportlab>=4.0.0
